+++ math/rotation_matrix.h
@@ -3,4 +3,4 @@
 // TODO(mitiguy) Merge Rotation class from
 // drake/multibody/multibody_tree/math/rotation.h into
 // drake/math/rotation_matrix.h.
-#include "drake/math/rotation_matrix_utilities.h"
+#include "drake/math/rotation_matrix_deprecated.h"
