+++ math/rotation_matrix_deprecated.cc
